Clean up infinite loop detection logic
Store what plugin was called along with the data it's called with.
If the same plugin is called with the same data twice, we know it will
loop infinitely.

As long as the data changes in each plugin run, the plugin should eventually
reach a result.

We may need to add further logic to prevent a stack overflow though.