Formalize returning data from a plugin
This isn't used by Swish directly, but will be passed on to other plugins.

Since plugin output is getting richer (MD, HTML etc.), it's becoming difficult to chain plugins together. Adding a data passing capability allows a plugin to process things and provide it in a usable way for another plugin to pick up and continue, rather than requiring the calling plugin to do weird parsing.