Update decode-jwt.ts
Move the 'header' to the bottom of the output - the payload is almost certainly more useful than the header.

Display the Issued At (iat), Not Before (nbf) and Expires (exp) properties as ISO8601 date times, and check if they are valid (iat/nbf in the past, exp in the future).